+++ CHANGES.rst
@@ -21,6 +21,7 @@ New Features
 
 - Add support for viewers with mixed pixel and wavelength/frequency spectral axis units. [#4001]
 
+- Table viewer tools to highlight selected, zoom-to-selected, and create subset from selected. [#3994]
 
 Cubeviz
 ^^^^^^^